fix: handle unexpected SELECT scenarios
This change makes the actuator ignore additional commands that may be
sent during the selection process. In particular, it ignores certain
GET DATA commands that are sent by the Windows Smart Card Plug and
Play Process. Additionally, SELECT commands referencing anything that
is not an AID (e.g. DO) are ignored.