Wait before disabling NFC in USER_DEVICE mode
Wait 1 second or until Reader is lost when disconnecting the transport
in USER_DEVICE mode, as some NFC messages may be waiting to be sent in
the NFC driver task.

This is required to make sure some messages are sent before the
transaction is completed, e.g. when sending EXCHANGE response as last
message of the transaction.

Signed-off-by: Adrian Gielniewski <adrian.gielniewski@nordicsemi.no>